requirements.txt
# Flask==2.3.2
# pymysql==1.1.0
# Flask==2.3.2 
# Flask-MySQLdb==1.0.1 
# bcrypt==4.0.1 
# gunicorn==20.1.0
# Flask==3.0.3 
# Flask-MySQLdb==2.0.0 
# bcrypt==4.2.0 
# gunicorn==23.0.0 
# python-dotenv==1.0.1
Flask==3.0.3 
Flask-MySQLdb==2.0.0 
mysqlclient==2.2.7 
bcrypt==4.2.0 
gunicorn==23.0.0 
python-dotenv==1.0.1